requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# API clients
tweepy>=4.4.0
python-telegram-bot>=20.3

# Analysis
pandas>=2.0.3
numpy>=1.21.0

# Testing
pytest>=6.2.5
